infrastructure/pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = "AWS CDK Infra"
authors = ["Lucio Tudisco <lucio.tudisco@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.118.0"
construct = "^2.10.70"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
